requirements.txt
flask
pillow
transparent-background
torch
torchvision
torchaudio --index-url https://download.pytorch.org/whl/cpu